feat: some cleaning
