+++ app/logger/logger.go
@@ -1,4 +1,4 @@
-package server
+package logger
 
 import (
 	"fmt"
