--- app/repository/postgres/dbmodel.go
@@ -1,9 +1,0 @@
-package postgres
-
-type DBPerson struct {
-	ID        int64
-	FirstName *string
-	LastName  *string
-	SurName   *string
-	Login     *string
-}
